+++ crates/puzzled_nono/src/puzzle/style/color.rs
@@ -1,2 +1,2 @@
-pub type ColorId = u16;
+pub type ColorId = usize;
 pub type Color = (u8, u8, u8);
