+++ crates/puzzled_nono_tui/src/app/pos.rs
@@ -3,14 +3,14 @@ use ratatui::layout::Position as AppPosition;
 
 pub fn puzzle_to_app(pos: PuzzlePosition) -> AppPosition {
     AppPosition {
-        x: pos.col,
-        y: pos.row,
+        x: pos.col as u16,
+        y: pos.row as u16,
     }
 }
 
 pub fn app_to_puzzle(pos: AppPosition) -> PuzzlePosition {
     PuzzlePosition {
-        row: pos.y,
-        col: pos.x,
+        row: pos.y as usize,
+        col: pos.x as usize,
     }
 }
